% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-calculate.R
\name{align_positive_poles}
\alias{align_positive_poles}
\title{Align all preferred poles on right side}
\usage{
align_positive_poles(x)
}
\description{
Align all preferred poles on right side
}
\keyword{internal}
