% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-utils.R
\name{alter_split}
\alias{alter_split}
\title{Spezieller String Split für Alter Definition}
\usage{
alter_split(x)
}
\arguments{
\item{x}{String zum splitten}
}
\description{
Alter in denen eine Maßnahme statfinden soll
können als in einem String als Abfolge eingegeben
werden.
}
\examples{
alter_split("10, 20,30" )
alter_split("; ,  10  ,,,  20;30,," ) # sehr robust

}
\keyword{internal}
