% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-utils.R
\name{cell_text_split}
\alias{cell_text_split}
\title{String splitter for comma separated values in Excel cell}
\usage{
cell_text_split(x)
}
\arguments{
\item{x}{String to be split}
}
\description{
String splitter for comma separated values in Excel cell
}
\examples{

cell_text_split("10, 20,30")
cell_text_split("; ,  10  ,,,  20;30,,") # very robust

}
\keyword{internal}
