% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-excel.R
\name{check_excel_input}
\alias{check_excel_input}
\title{Check if Excel input file contains valid data}
\usage{
check_excel_input(x)
}
\arguments{
\item{x}{Data from Excel input file.}
}
\description{
Check if Excel input file contains valid data
}
