% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-calculate.R
\name{clique_color_pals}
\alias{clique_color_pals}
\title{Generate colors for cliques}
\usage{
clique_color_pals(n, name = "Dark2", alpha = 0.1)
}
\arguments{
\item{n}{Number of colors.}

\item{name}{Name of RColorBrewer qualitative palette.}

\item{aplha}{Alpha color value for fill colors.}
}
\value{
A list with a vector of border and fill colors.
}
\description{
Generate colors for cliques
}
\keyword{internal}
