% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-excel.R
\name{create_excel_output}
\alias{create_excel_output}
\title{Create output Excel file}
\usage{
create_excel_output(file, data = list())
}
\arguments{
\item{file}{Path to workbook.}

\item{data}{Named list of data objects to add to Excel file.
The following contents are expected: TODO}
}
\value{
Path to saved file.
}
\description{
Loads the supplied workbook and adds calculations
}
