% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-utils.R
\name{dt_default}
\alias{dt_default}
\title{Default datatable output when loading}
\usage{
dt_default(text = "Waiting for data ...", header = "")
}
\arguments{
\item{text}{Text to be shown}
}
\description{
Default datatable output when loading
}
