% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-utils.R
\name{emptify_object}
\alias{emptify_object}
\title{Create empty dataframe while keeping columns names}
\usage{
emptify_object(x)
}
\arguments{
\item{x}{Dataframe or list.}
}
\description{
Also works on lists if dataframes are contained as list elements.
}
\keyword{internal}
