% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-utils.R
\name{make_names_vec}
\alias{make_names_vec}
\alias{make_names}
\title{standardisierte variablen namen}
\usage{
make_names_vec(x)

make_names(x)
}
\arguments{
\item{x}{Dataframe oder Vektor}
}
\description{
standardisierte variablen namen
}
\keyword{internal}
