% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OpenRepGrid.ic.R
\docType{package}
\name{OpenRepGrid.ic-package}
\alias{OpenRepGrid.ic-package}
\title{\pkg{OpenRepGrid.ic} - OpenRepGrid.ic}
\description{
\pkg{OpenRepGrid.ic} - OpenRepGrid.ic
}
\keyword{OpenRepGrid.ic}
\keyword{package}
