% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-calculate.R
\name{valence_pole_prefix}
\alias{valence_pole_prefix}
\title{Prefix the pole label with +/- for indicate pole valence}
\usage{
valence_pole_prefix(x, valence, prefix = "(", postfix = ")")
}
\description{
Prefix the pole label with +/- for indicate pole valence
}
\keyword{internal}
