% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-calculate.R
\name{calculate_similarity}
\alias{calculate_similarity}
\title{Calculate similarity matrix}
\usage{
calculate_similarity(x, min_matches = 6, align_poles = TRUE)
}
\arguments{
\item{x}{Grid data.}

\item{min_matches}{Minimal number of matches to considers constructs as related.}

\item{align_poles}{Align positive poles on the right and negative poles on the left.}
}
\description{
Calculate similarity matrix
}
