% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OpenRepGrid.ic.R
\name{ic}
\alias{ic}
\title{Launch app in browser}
\usage{
ic(display.mode = "auto", launch.browser = TRUE)
}
\arguments{
\item{display.mode}{\code{auto} by default, can also be \code{showcase}. See
\link[shiny]{runApp}.}

\item{launch.browser}{Boolean, set \code{TRUE} (default) to open the app in
the browser. See \link[shiny]{runApp}.}
}
\description{
Launch app in browser
}
\examples{
if (interactive()) {
  ic()
}
}
