% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-utils.R
\name{make_names_vec}
\alias{make_names_vec}
\alias{make_names}
\title{standardized variable names}
\usage{
make_names_vec(x)

make_names(x)
}
\arguments{
\item{x}{Dataframe or vector}
}
\description{
standardized variable names
}
\keyword{internal}
